refactor UploadImages
