+++ components/ImageUpload.tsx
@@ -1,14 +1,79 @@
 'use client'
 
-import { Suspense } from 'react'
+import { Suspense, useCallback, useEffect, useState } from 'react'
 import { X, UploadCloud } from 'lucide-react'
 import { cn } from '@/lib/utils'
 import { SkeletonGrid } from './skeletons/SkeletonGridImages'
 import Image from 'next/image'
-import useImageUpload from '@/hooks/useImageUpload'
+import { FileRejection, useDropzone } from 'react-dropzone'
+import { useToastHandler } from '@/hooks/useToastHandler'
+import { FilesType } from '@/types'
 
-export default function ImageUploader() {
-  const { files, getInputProps, getRootProps, isDragActive, removeFile } = useImageUpload()
+type ImageUploaderProps = {
+  files: FilesType[]
+  setFiles: React.Dispatch<React.SetStateAction<FilesType[]>>
+}
+
+export default function ImageUploader({ files, setFiles }: ImageUploaderProps) {
+  const { showErrorToast, showPreventiveToast } = useToastHandler()
+
+  useEffect(() => {
+    return () => files.forEach(file => URL.revokeObjectURL(file.preview))
+  }, [files])
+
+  const onDrop = useCallback(
+    (acceptedFiles: File[], rejectedFiles: FileRejection[]) => {
+      if (acceptedFiles?.length) {
+        const totalFiles = files.length + acceptedFiles.length
+        if (totalFiles > 4) {
+          showPreventiveToast('Se ha excedido el límite de 4 archivos. Solo se aceptarán los primeros archivos hasta completar el límite.')
+
+          const remainingSlots = Math.max(0, 4 - files.length)
+          const allowedFiles = acceptedFiles.slice(0, remainingSlots)
+
+          if (remainingSlots > 0) {
+            const mappedFiles: FilesType[] = allowedFiles.map(file => ({
+              ...file,
+              path: file.name,
+              relativePath: file.name,
+              preview: URL.createObjectURL(file)
+            }))
+            setFiles(prev => [...prev, ...mappedFiles])
+          }
+        } else {
+          const mappedFiles: FilesType[] = acceptedFiles.map(file => ({
+            ...file,
+            path: file.name,
+            relativePath: file.name,
+            preview: URL.createObjectURL(file)
+          }))
+          setFiles(prev => [...prev, ...mappedFiles])
+        }
+      }
+
+      if (rejectedFiles?.length) {
+        try {
+          rejectedFiles.forEach(file => {
+            const errorMessage = file.errors?.map(error => error.message)?.join(', ') ?? 'Error desconocido'
+            showErrorToast(errorMessage)
+          })
+        } catch {
+          showErrorToast('Error al procesar los archivos rechazados')
+        }
+      }
+    },
+    [files, showPreventiveToast, showErrorToast]
+  )
+
+  const { getRootProps, getInputProps, isDragActive } = useDropzone({
+    accept: { 'image/*': [] },
+    onDrop,
+    maxSize: 5242880 // 5MB
+  })
+
+  const removeFile = (path: string) => {
+    setFiles(prev => prev.filter(file => file.path !== path))
+  }
 
   return (
     <section>
@@ -28,11 +93,12 @@ export default function ImageUploader() {
               ? 'Suelta las imágenes aquí'
               : files.length >= 4
               ? 'Límite de imágenes alcanzado'
-              : 'Arrastra y suelta imágenes aquí, o haz click para seleccionar - maximo 4 imagenes'}
+              : 'Arrastra y suelta imágenes aquí, o haz click para seleccionar - máximo 4 imágenes'}
           </p>
-          <p className='text-xs text-gray-500 text-center '>PNG, JPG, WEBP, AVIF, SVG - Max 5MB</p>
+          <p className='text-xs text-gray-500 text-center'>PNG, JPG, WEBP, AVIF, SVG - Máx 5MB</p>
         </div>
       </div>
+
       <Suspense fallback={<SkeletonGrid count={files.length} />}>
         <ul className={files.length ? 'my-10 grid grid-cols-2 sm:gap-y-20 lg:grid-cols-4 gap-10' : 'hidden'}>
           {files.map(file => (
